backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
opencv-python-headless
numpy
python-dotenv
httpx
python-multipart
pillow
pillow-heif